whatsapp_pip_requirement.txt
aniso8601==1.2.0
Appium-Python-Client==0.26
asn1crypto==0.23.0
base64-util==0.1
bcrypt==3.1.3
bitmath==1.3.1.2
blinker==1.4
cachez==0.1.2
certifi==2017.4.17
cffi==1.11.0
chardet==3.0.2
click==6.7
cryptography==2.0.3
ddtrace==0.11.1
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
flasgger==0.6.4
Flask==0.12
Flask-HTTPAuth==2.2.0
Flask-Mail==0.9.1
Flask-Prometheus==0.0.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
functools32==3.2.3.post2
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
idna==2.6
importlib==1.0.4
influxdb==4.1.0
ipaddress==1.0.18
IPy==0.83
itsdangerous==0.23
Jinja2==2.9.5
jsonschema==2.6.0
lxml==3.8.0
MarkupSafe==1.0
meld3==1.0.2
mistune==0.7.4
msgpack-python==0.5.6
multi-key-dict==2.0.3
MySQL-python==1.2.5
paramiko==2.3.1
passlib==1.6.1
pbr==2.0.0
persist-queue==0.3.4
phonenumbers==8.9.9
pika==0.11.2
pluggy==0.4.0
prometheus-client==0.1.0
py==1.4.33
pyasn1==0.3.6
PyChef==0.3.0
pycparser==2.18
pydns==2.3.6
PyMySQL==0.7.10
PyNaCl==1.1.2
python-3parclient==4.2.4
python-dateutil==2.6.0
python-jenkins==0.4.14
pytz==2016.10
PyYAML==3.12
requests==2.13.0
requests-futures==0.9.5
retryz==0.1.9
selenium==3.13.0
six==1.10.0
SQLAlchemy==1.1.7
supervisor==3.3.1
tox==2.7.0
utils==0.9.0
uWSGI==2.0.17
validate-email==1.3
virtualenv==15.1.0
Werkzeug==0.12.1
wrapt==1.10.11